pyproject.toml
[tool.poetry]
name = "hotcities"
version = "0.1.0"
description = ""
authors = ["Flavio Grandin"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.4.1"
numpy = "^1.25.1"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
click = "^8.1.5"
google-api-python-client = "^2.94.0"
google-cloud-vision = "^3.4.4"
requests = "^2.31.0"
flickrapi = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
poethepoet = "^0.21.1"
ipykernel = "^6.24.0"
jupyter = "^1.0.0"
pylint = "^2.17.4"

[tool.poetry.scripts]
hotcities = "hotcities.cli:cli"

[tool.poe.tasks]
lint = "pylint hotcities"
test = "pytest test"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

